Better testing, more dyamnic score updates.

+++ bot.py
@@ -20,18 +20,19 @@ for x in open("info", "r").read().split("\n"):
         info.append(x)
 
 # Info
-prefix   = info[0]
-status   = info[1]
+prefix = info[0]
+status = info[1]
+state  = info[2]
 
 # Channels
-CommandChannel = info[2]
-UpdateChannel  = info[3]
+CommandChannel = info[3]
+UpdateChannel  = info[4]
 
 # Credentials
-DiscordKey   = info[4]
-CTFDUsername = info[5]
-CTFDPassword = info[6]
-CTFDURL      = info[7]
+DiscordKey   = info[5]
+CTFDUsername = info[6]
+CTFDPassword = info[7]
+CTFDURL      = info[8]
 
 # API
 bot = discord.Client()
@@ -82,7 +83,6 @@ async def on_message(message):
 
 @bot.event
 async def on_ready():
-    print("Bot Running.")
     await bot.change_presence(activity = discord.Game(name = status + " Prefix : " + prefix))
 
 ## Api Handling ##
@@ -191,11 +191,12 @@ def createTeam(name : str):
     data = {
         "name" : name,
     }
-    res = json.loads(api.post("https://playcodecup.com/api/v1/teams", json = data, headers = {'Content-Type':'application/json'}).text)
+    if state.lower() == "active" or state.lower() == "api":
+        res = json.loads(api.post("https://playcodecup.com/api/v1/teams", json = data, headers = {'Content-Type':'application/json'}).text)
     try:
-        if res["success"]:
+        if res and res["success"]:
             return "Created Team : `" + name + "`."
-    except KeyError:
+    except:
         return "Unable to create Team."
 
 def removeTeam(name : str):
@@ -206,14 +207,18 @@ def removeTeam(name : str):
             teamid = i["id"]
     if not teamid:
         return "No team with name : `" + name + "` exists." 
-    res = json.loads(api.delete("https://playcodecup.com/api/v1/teams/" +  str(teamid), headers = {'Content-Type':'application/json'}).text)
+    if state.lower() == "active" or state.lower() == "api":
+        res = json.loads(api.delete("https://playcodecup.com/api/v1/teams/" +  str(teamid), headers = {'Content-Type':'application/json'}).text)
     try:
-        if res["success"]:
+        if res and res["success"]:
             return "Removed Team : `" + name + "`."
-    except KeyError:
+    except:
         return "Unable to remove Team."
 
-def updateSolves():
+# Run
+
+def runDymanic():
+    print("Running Dymanic Scores")
     global challenges, questions
     while True:
         time.sleep(3)
@@ -223,20 +228,21 @@ def updateSolves():
 
         solves = {}
         for i in questions:
-            solves[i["id"]] = getSolves(i["id"])
+            solves[i["id"]] = json.loads(api.get("https://playcodecup.com/api/v1/challenges/" + str(i["id"]) + "/solves").text)["data"]
         
         solves = sorted(solves.items(), key = lambda x : x[1])
         zeros  = list(filter(None, (x[0] if x[1] == 0 else None for x in solves)))
-        print(zeros)
 
-        upgrade = random.choice(zeros)
-        for i in questions:
-            if i["id"] == upgrade:
-                upgrade = i
-                break
-        
+        if len(zeros) != 0:
+            upgrade = random.choice(zeros)
+            for i in questions:
+                if i["id"] == upgrade:
+                    upgrade = i
+                    break
+        else:
+            upgrade = solves[-1]
+
         if type(upgrade) == int: continue
-        print(upgrade)
         upgrade = json.loads(api.get("https://playcodecup.com/api/v1/challenges/" + str(upgrade["id"])).text)["data"]
         data = {
             "name": upgrade['name'],
@@ -246,10 +252,46 @@ def updateSolves():
             "max_attempts": upgrade["max_attempts"],
             "state": upgrade['state']
         }
-        print("https://playcodecup.com/api/v1/challenges/" + str(upgrade["id"]))
-        Res = api.patch("https://playcodecup.com/api/v1/challenges/" + str(upgrade["id"]), json = data, headers = {'Content-Type':'application/json'})
-        print(Res.text)
+        if state.lower() == "active" or state.lower() == "api":
+            Res = api.patch("https://playcodecup.com/api/v1/challenges/" + str(upgrade["id"]), json = data, headers = {'Content-Type':'application/json'})
 
-# Run
-#bot.run(DiscordKey)
-#threading.Thread(target = UpdateSolves).start()
+def runDiscord():
+    print("Running Discord Bot")
+    bot.run(DiscordKey)
+
+def Run():
+    global state
+    print("Running : codecup-discord-help")
+    print("State : " + state)
+    if state.lower() == "active" or state.lower() == "discord" or state.lower() == "passive":
+        threading.Thread(target = runDiscord).start()
+    if state.lower() == "active" or state.lower() == "api" or state.lower() == "passive":
+        threading.Thread(target = runDymanic).start()
+    if state.lower() == "testing":
+        c = ""
+        while c != "quit":
+            c = input(" > ")
+            if c == "active":
+                state = "Active"
+                threading.Thread(target = runDiscord).start()
+                threading.Thread(target = runDymanic).start()
+            elif c == "passive":
+                state = "Passive"
+                print("State set to : Passive")
+            elif c == "api":
+                state = "API"
+                print("State set to : API")
+            elif c == "discord":
+                state = "Discord"
+                print("State set to : Discord")
+            elif c == "dymanic":
+                threading.Thread(target = runDymanic).start()
+            elif c == "bot":
+                threading.Thread(target = runDiscord).start()
+            else:
+                try:
+                    print(eval(c))
+                except Exception as e:
+                    print(e)
+            
+Run()
